+++ wrapper/include/descriptors.h
@@ -12,6 +12,6 @@ namespace RDKit {
     std::unique_ptr<std::vector<std::string>> get_property_names(std::shared_ptr<Properties> props);
     std::unique_ptr<std::vector<double>> compute_properties(std::shared_ptr<Properties> props, std::shared_ptr<ROMol> mol);
     rust::String mol_formula(std::shared_ptr<ROMol> mol);
-    rust::i32 symmetrize_SSSR(std::shared_ptr<ROMol> mol);
-    rust::i32 mol_exact_MW(std::shared_ptr<ROMol> mol);
+    rust::u16 symmetrize_SSSR(std::shared_ptr<ROMol> mol);
+    rust::u32 mol_exact_MW(std::shared_ptr<ROMol> mol);
 }
